requirements.txt
flask
flask-sqlalchemy
flask-cors
flask-limiter
feedparser
requests
beautifulsoup4
lxml
Pillow
numpy
apscheduler
python-dotenv
gunicorn
